Start refactoring compress index map
