+++ cpp/dolfinx/common/IndexMap.cpp
@@ -153,6 +153,196 @@ compute_owned_shared(MPI_Comm comm, const xtl::span<const std::int64_t>& ghosts,
 } // namespace
 
 //-----------------------------------------------------------------------------
+std::pair<IndexMap, std::vector<std::int64_t>>
+common::compress_index_map(const IndexMap& map,
+                           const xtl::span<const std::int32_t>& indices)
+{
+  // Compute number of owned indices in the new map
+  auto it = std::lower_bound(indices.begin(), indices.end(), map.size_local());
+  std::int64_t local_size_new = std::distance(indices.begin(), it);
+
+  MPI_Comm comm = map.comm(dolfinx::common::IndexMap::Direction::reverse);
+
+  // Compute global offset (index), using partial exclusive reduction
+  std::int64_t offset_new = 0;
+  MPI_Request request_scan;
+  MPI_Iexscan(&local_size_new, &offset_new, 1, MPI_INT64_T, MPI_SUM, comm,
+              &request_scan);
+  MPI_Wait(&request_scan, MPI_STATUS_IGNORE);
+
+  // New local range
+  // const std::array<std::int64_t, 2> local_range
+  //     = {offset_new, offset_new + local_size_new};
+
+  // Send local size to sum reduction to get global size
+  std::int64_t size_global = 0;
+  MPI_Request request;
+  MPI_Iallreduce(&local_size_new, &size_global, 1, MPI_INT64_T, MPI_SUM, comm,
+                 &request);
+  // Wait for the MPI_Iallreduce to complete
+  MPI_Wait(&request, MPI_STATUS_IGNORE);
+
+  // Build array of global indices for the owned indices in the new map
+  std::vector<std::int64_t> global_indices_new(map.size_local(), -1);
+  for (std::size_t i = 0; i < indices.size(); ++i)
+    global_indices_new[indices[i]] = i + offset_new;
+
+  // TODO: Use scatter_fwd_begin/end for efficiency
+
+  // Send my new global index to ranks that ghost my indices
+  std::vector<std::int64_t> my_ghosts_indices(map.num_ghosts());
+  map.scatter_fwd(xtl::span<const std::int64_t>(global_indices_new),
+                  xtl::span<std::int64_t>(my_ghosts_indices), 1);
+
+  std::size_t num_ghosts_new = std::distance(it, indices.end());
+  std::vector<std::int64_t> ghosts_new;
+  ghosts_new.reserve(num_ghosts_new);
+  for (auto it1 = it; it1 != indices.end(); ++it1)
+    ghosts_new.push_back(my_ghosts_indices[*it1]);
+
+  // TODO: Update communication data and create new index map
+
+
+  // ---- old
+
+  const std::vector<std::int64_t>& ghosts = map.ghosts();
+
+  std::vector<int> ghost_owners = map.ghost_owner_rank();
+
+  // Commmunicate ghost indices
+  MPI_Comm reverse_comm
+      = map.comm(dolfinx::common::IndexMap::Direction::reverse);
+  MPI_Comm forward_comm
+      = map.comm(dolfinx::common::IndexMap::Direction::forward);
+  auto [source_ranks, dest_ranks] = dolfinx::MPI::neighbors(forward_comm);
+
+  // Create inverse map of src ranks
+  std::unordered_map<std::int32_t, std::int32_t> rank_glob_to_loc;
+  for (std::size_t i = 0; i < source_ranks.size(); ++i)
+    rank_glob_to_loc.insert({source_ranks[i], i});
+  // Split indices into local and ghost indices
+  const std::int32_t local_size = map.size_local();
+  std::vector<std::int32_t> local_indices, ghost_indices;
+  const std::int32_t num_local
+      = std::count_if(indices.begin(), indices.end(),
+                      [local_size](int i) { return i < local_size; });
+  local_indices.reserve(num_local);
+  ghost_indices.reserve(indices.size() - num_local);
+
+  // Counter of number of ghosts per process
+  std::vector<std::int32_t> num_ghosts(source_ranks.size());
+
+  // Split input indices into local and ghost entities, and count
+  // how many to send to each owner
+  for (std::size_t i = 0; i < indices.size(); ++i)
+  {
+    if (indices[i] < local_size)
+      local_indices.push_back(indices[i]);
+    else
+    {
+      const std::int32_t ghost_index = indices[i] - local_size;
+      ghost_indices.push_back(ghost_index);
+      const std::int32_t ghost_rank = ghost_owners[ghost_index];
+      num_ghosts[rank_glob_to_loc[ghost_rank]]++;
+    }
+  }
+
+  // Compute insertion position in adjacency array
+  std::vector<std::int32_t> ghost_adj_offset(source_ranks.size() + 1, 0);
+  std::partial_sum(num_ghosts.begin(), num_ghosts.end(),
+                   ghost_adj_offset.begin() + 1);
+  std::vector<std::int64_t> ghost_adj(ghost_adj_offset.back());
+
+  std::vector<std::int32_t> inserted_ghosts(source_ranks.size(), 0);
+  for (std::size_t i = 0; i < ghost_indices.size(); ++i)
+  {
+    const std::int32_t ghost_index = ghost_indices[i];
+    const std::int32_t rank = rank_glob_to_loc[ghost_owners[ghost_index]];
+    const std::int32_t local_pos
+        = ghost_adj_offset[rank] + inserted_ghosts[rank];
+    ghost_adj[local_pos] = ghosts[ghost_index];
+    inserted_ghosts[rank]++;
+  }
+  dolfinx::graph::AdjacencyList<std::int64_t> send_ghosts(ghost_adj,
+                                                          ghost_adj_offset);
+
+  // Send ghosts to owner
+  dolfinx::graph::AdjacencyList<std::int64_t> recv_ghosts
+      = dolfinx::MPI::neighbor_all_to_all(reverse_comm, send_ghosts);
+
+  // Compute new destination ranks for received ghosts
+  std::vector<std::int32_t> new_dest_ranks;
+  for (std::int32_t i = 0; i < recv_ghosts.num_nodes(); ++i)
+    if (recv_ghosts.num_links(i) > 0)
+      new_dest_ranks.push_back(dest_ranks[i]);
+
+  // Convert local input indices to global indices
+  std::vector<std::int64_t> global_indices(local_indices.size());
+  map.local_to_global(local_indices, global_indices);
+
+  // Create a sorted vector of indices owned by the process by merging
+  // input indices with received ghost indices, removing duplicates
+  std::vector<std::int64_t> sorted_ghosts = recv_ghosts.array();
+  std::sort(sorted_ghosts.begin(), sorted_ghosts.end());
+  std::sort(global_indices.begin(), global_indices.end());
+  std::vector<std::int64_t> org_global_indices;
+  org_global_indices.reserve(local_indices.size() + sorted_ghosts.size());
+  std::merge(global_indices.begin(), global_indices.end(),
+             sorted_ghosts.begin(), sorted_ghosts.end(),
+             std::back_inserter(org_global_indices));
+  org_global_indices.erase(
+      std::unique(org_global_indices.begin(), org_global_indices.end()),
+      org_global_indices.end());
+
+  // Compute global indices for new index map
+  std::size_t offset = 0;
+  std::size_t new_num_local = org_global_indices.size();
+  MPI_Exscan(&new_num_local, &offset, 1, dolfinx::MPI::mpi_type<std::size_t>(),
+             MPI_SUM, MPI_COMM_WORLD);
+
+  // For each received ghost, compute new global index
+  const std::vector<std::int64_t>& ghost_array = recv_ghosts.array();
+  std::vector<std::int64_t> ghosts_map(2 * ghost_array.size());
+  for (std::size_t i = 0; i < ghost_array.size(); ++i)
+  {
+    auto idx_it = std::lower_bound(org_global_indices.begin(),
+                                   org_global_indices.end(), ghost_array[i]);
+    std::int32_t index = std::distance(org_global_indices.begin(), idx_it);
+    ghosts_map[2 * i] = ghost_array[i];
+    ghosts_map[2 * i + 1] = index + offset;
+  }
+
+  // Send map of indices
+  std::vector<std::int32_t> send_offsets = recv_ghosts.offsets();
+  for (std::size_t i = 0; i < send_offsets.size(); ++i)
+    send_offsets[i] *= 2;
+  dolfinx::graph::AdjacencyList<std::int64_t> send_new_ghosts(ghosts_map,
+                                                              send_offsets);
+  dolfinx::graph::AdjacencyList<std::int64_t> new_ghosts_adj
+      = dolfinx::MPI::neighbor_all_to_all(forward_comm, send_new_ghosts);
+  // Get new ghosts and src ranks
+  std::vector<std::int64_t> new_ghosts;
+  new_ghosts.reserve(ghost_indices.size());
+  std::vector<std::int32_t> new_src_ranks;
+  new_src_ranks.reserve(ghost_indices.size());
+
+  for (std::int32_t i = 0; i < new_ghosts_adj.num_nodes(); ++i)
+  {
+    const std::int32_t src_rank = source_ranks[i];
+    auto ghost_data = new_ghosts_adj.links(i);
+    for (std::size_t j = 0; j < ghost_data.size() / 2; ++j)
+    {
+      org_global_indices.push_back(ghost_data[2 * j]);
+      new_ghosts.push_back(ghost_data[2 * j + 1]);
+      new_src_ranks.push_back(src_rank);
+    }
+  }
+
+  return {IndexMap(reverse_comm, new_num_local, new_dest_ranks, new_ghosts,
+                   new_src_ranks),
+          std::move(org_global_indices)};
+}
+//-----------------------------------------------------------------------------
 std::tuple<std::int64_t, std::vector<std::int32_t>,
            std::vector<std::vector<std::int64_t>>,
            std::vector<std::vector<int>>>
